- adds Cmdb and CmConfig types/fields to the Hpcm struct - adds functions for setting up an hpcm client - converges existing code into a Translate function for better   readability/portability - adds/updates all required provider commands - adds functions for extracting models, vendors, and other metadata from   variuous hpcm_client types - adds functions to extract geolocation data when possible - adds function to get convert to a cani hardwaretype - import from Cmdb functions/gather all hardware
