+++ internal/provider/hpcm/export.go
@@ -2,7 +2,7 @@
  *
  *  MIT License
  *
- *  (C) Copyright 2023 Hewlett Packard Enterprise Development LP
+ *  (C) Copyright 2023-2024 Hewlett Packard Enterprise Development LP
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a
  *  copy of this software and associated documentation files (the "Software"),
@@ -32,6 +32,6 @@ import (
 )
 
 func (hpcm *Hpcm) Export(*cobra.Command, []string, inventory.Datastore) error {
-	log.Warn().Msgf("not yet implemented")
+	log.Warn().Msgf("Export not yet implemented")
 	return nil
 }
